Basement Textures and Table Change-Collisions Need Adding
Most of the textures have been properly applied now with proper unwraps. 

A model update I made on the table just to improve visual fidelity. Size remains the same.

Branching as I deleted the original model that had collisions and didn't want to break anything that was in the scene. Had to due to original models not having correct UVs and would have been hard to replace every item in the scene.