All Textures Applied to Basement
All textures applied.

No new models will be added to the scene. 

Need feedback on the board design. Feel that it is an improvement on the last board model that was made by yamil. Thought it could immerse the player into the world a bit more. The board is currently just the  terrain from one of the other levels. This can be improved on if we like the new design and want to see it look more tabletop/board gameish. The terrain elevation could be more subtle for example.

Level Defence One still needs alot of work and another level needs to be made. Either with existing assets made by Yamil or some new ones that i could make just to make each level a more unique feel.