Lighting For Basement & Level 1 Tower Scene
Added lighting for basement and Level 1.

Level 1 lighting could do with tweaking since the fog/mist effect might be too strong. Wanted it to be more atmospheric in the scene and so that it is harder to see where the enemies are coming from/ spawning in. Drop feedback in discord for what it looks like in VR. 

Working in 2nd Level models and lighting and as well as figuring something out for level 3 since we still dont have anything